// client/src/api/model/campaignEncounterEncounterStatus.ts
/**
 * Generated by orval v8.0.0 🍺
 * Do not edit manually.
 * Genesys API Documentation
 * Interactive API documentation
 * OpenAPI spec version: 1.0
 */

export type CampaignEncounterEncounterStatus = typeof CampaignEncounterEncounterStatus[keyof typeof CampaignEncounterEncounterStatus];


export const CampaignEncounterEncounterStatus = {
  Ready: 'Ready',
  Active: 'Active',
  Resolved: 'Resolved',
} as const;