// client/src/api/model/campaignEncounterEncounterType.ts
/**
 * Generated by orval v7.19.0 🍺
 * Do not edit manually.
 * Genesys API Documentation
 * Interactive API documentation
 * OpenAPI spec version: 1.0
 */

export type CampaignEncounterEncounterType = typeof CampaignEncounterEncounterType[keyof typeof CampaignEncounterEncounterType];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const CampaignEncounterEncounterType = {
  Combat: 'Combat',
  Social: 'Social',
  General: 'General',
} as const;