// client/src/api/model/campaignSessionStatus.ts
/**
 * Generated by orval v8.4.2 🍺
 * Do not edit manually.
 * Genesys API Documentation
 * Interactive API documentation
 * OpenAPI spec version: 1.0
 */

export type CampaignSessionStatus = typeof CampaignSessionStatus[keyof typeof CampaignSessionStatus];


export const CampaignSessionStatus = {
  Planning: 'Planning',
  Ready: 'Ready',
  Active: 'Active',
  Resolved: 'Resolved',
} as const;